Fix axios mishandling TypedArrays with buffer pools
Protobuf.js encodes protobuf messages with an 8kB buffer pool,
which axios will mishandle if encoded as a Uint8Array. As a result,
when trying to submit a payload from the browser (but not from Node)
8kB of mostly empty bytes will be submitted to REST API and rejected.

This fixes the problem with a fairly straightford fix, submitting a
slice of the Uint8Array, which will create a new array with a buffer
of the same length as the array, rather than 8kb.

Signed-off-by: Zac Delventhal <zac@bitwise.io>