Clarify usage of randomBytes in signing tests
These random messages have confused multiple students into thinking there is
more going on in the tests than there is. The usage has been tweaked with
some explanatory comments added to make it more clear.

There is also a new note in the README explaining when strings will be signed,
and when Buffers.

Fixes #88

Signed-off-by: Zac Delventhal <zac@bitwise.io>