Refactor promiseless error workaround
The release of Sawtooth 1.0.5 has been delayed, which necessitates a more
robust workaround for the bug caused by throwing errors outside of a Promise.
It is much cleaner to use a closure to wrap the handler's apply method
in a try/catch than to add it to the method itself. This has the added
benefit of being implementable in the index.js code which is provided
to students.

Signed-off-by: Zac Delventhal <zac@bitwise.io>